scheduler_app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
alembic==1.16.5
python-decouple==3.8
requests==2.31.0
PyPDF2==3.0.1
reportlab==4.0.7
urllib3==2.0.7
celery==5.3.4
redis==5.0.1
cryptography>=41.0.0
xhtml2pdf>=0.2.11
pytest>=7.4.3
python-barcode>=0.15.1
Pillow>=10.0.0